testing/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
pyserial>=3.5
argparse 